Split endpoint helpers into focused modules
